Add ExecutingBuildOperation
And use it to track input snapshotting and build cache key calculation.
