+++ subprojects/base-services/src/main/java/org/gradle/internal/operations/BuildOperationExecutor.java
@@ -51,6 +51,13 @@ public interface BuildOperationExecutor {
     <T> T call(CallableBuildOperation<T> buildOperation);
 
     /**
+     * Starts an operation that can be finished later through its context available via {@link ExecutingBuildOperation#getContext()}.
+     *
+     * When a parent operation is finished any unfinished child operations will be failed.
+     */
+    ExecutingBuildOperation start(BuildOperationDescriptor.Builder descriptor);
+
+    /**
      * Submits an arbitrary number of runnable operations, created synchronously by the scheduling action, to be executed in the global
      * build operation thread pool. Operations may execute concurrently. Blocks until all operations are complete.
      */
